11/27/18 Objective
linked stylesheets in articles. Fixed styles in Flex. Added a nav-icon so in mobile the navigation list is not taking up as much of the website. Linked nav-icon css to prek. Created a stylesheet so when the website is printed it doesn't have more then necessary information, like hiding the nav bar or displaying url instead of links.